Add retrieval of neighbouring support points for DCPSE.
The Support class allow getting the required number of neighbouring points of a given one.
It does so in a way that only expands the search area in directions where we are performing derivation.
Now linking to a newer version of openfpm_data as a fix in const-ness of some operators was required.
Also minor fixes to some other tests.