Add RHS constuction for DCPSE and fix build issue when EIGEN package not properly found at first.
The DcpseRhs class provides machinery for getting the RHS to be used in the linear system associated with DCPSE computation. Analogously to Vandermonde, the user has to initialize the class with the MonomialBasis and differential signature in use and then pass an existing vector object to be filled to the getVector() method.
The fix in CMakeLists.txt is to allow MacOs and CentOS builds to be successful: in these cases the EIGEN package may not be found at first and this case is now supported.